+++ src/collective/volto/formsupport/captcha/honeypot.py
@@ -1,8 +1,10 @@
 from . import CaptchaSupport
-from collective.honeypot.config import EXTRA_PROTECTED_ACTIONS
 from collective.honeypot.config import HONEYPOT_FIELD
-from collective.honeypot.utils import check_post
+from collective.honeypot.utils import found_honeypot
 from collective.volto.formsupport import _
+from plone.restapi.deserializer import json_body
+from zExceptions import BadRequest
+from zope.i18n import translate
 
 
 class HoneypotSupport(CaptchaSupport):
@@ -12,7 +14,7 @@ class HoneypotSupport(CaptchaSupport):
         """
         Honeypot is enabled with env vars
         """
-        return "submit-form" in EXTRA_PROTECTED_ACTIONS
+        return True
 
     def serialize(self):
         if not HONEYPOT_FIELD:
@@ -22,4 +24,12 @@ class HoneypotSupport(CaptchaSupport):
         return {"id": HONEYPOT_FIELD}
 
     def verify(self, data):
-        check_post()
+        form_data = json_body(self.request).get("data", [])
+        form = {x["label"]: x["value"] for x in form_data}
+        if found_honeypot(form, required=True):
+            raise BadRequest(
+                translate(
+                    _("honeypot_error", default="Error submitting form."),
+                    context=self.request,
+                )
+            )